refine trace
