+++ libs/system/trace/util.go
@@ -1,46 +1,109 @@
 package trace
 
 import (
-	"runtime"
-	"strings"
+	"github.com/spf13/viper"
 	"sync"
-	"time"
 )
 
-func RunFuncName() string {
-	pc := make([]uintptr, 1)
-	runtime.Callers(2, pc)
-	f := runtime.FuncForPC(pc[0])
-	names := strings.Split(f.Name(), ".")
-	return names[len(names)-1]
+const FlagEnableAnalyzer string = "enable-analyzer"
+
+var (
+	analyzer *Analyzer = &Analyzer{}
+	openAnalyzer bool
+	dynamicConfig IDynamicConfig = MockDynamicConfig{}
+	forceAnalyzerTags map[string]struct{}
+	status            bool
+	once              sync.Once
+)
+
+func EnableAnalyzer(flag bool)  {
+	status = flag
 }
 
-func GetNowTimeMs() int64 {
-	return time.Now().UnixNano() / 1e6
-}
-
-type DbRecord struct {
-	lock sync.RWMutex
-	oper map[string]int
-}
-
-func newDbRecord() *DbRecord {
-	return &DbRecord{
-		oper: make(map[string]int),
+func initForceAnalyzerTags() {
+	forceAnalyzerTags = map[string]struct{}{
+		RunAnte: {},
+		Refund:  {},
+		RunMsg:  {},
 	}
 }
 
-func (s *DbRecord) GetOperType(oper string) int {
-	s.lock.RLock()
-	defer s.lock.RUnlock()
-	if _, ok := s.oper[oper]; !ok {
-		return -1
+func init() {
+	initForceAnalyzerTags()
+
+	dbOper = newDbRecord()
+	for _, v := range STATEDB_READ {
+		dbOper.AddOperType(v, READ)
 	}
-	return s.oper[oper]
+	for _, v := range STATEDB_WRITE {
+		dbOper.AddOperType(v, WRITE)
+	}
+	for _, v := range EVM_OPER {
+		dbOper.AddOperType(v, EVMALL)
+	}
 }
 
-func (s *DbRecord) AddOperType(oper string, value int) {
-	s.lock.Lock()
-	defer s.lock.Unlock()
-	s.oper[oper] = value
+func OnAppBeginBlockEnter(height int64) {
+	newAnalyzer(height)
+	if !dynamicConfig.GetEnableAnalyzer() {
+		openAnalyzer = false
+		return
+	}
+	openAnalyzer = true
+	lastElapsedTime := GetElapsedInfo().GetElapsedTime()
+	if singlePprofDumper != nil && lastElapsedTime > singlePprofDumper.triggerAbciElapsed {
+		singlePprofDumper.cpuProfile(height)
+	}
+}
+
+func skip(a *Analyzer, oper string) bool {
+	if a != nil {
+		if openAnalyzer {
+			return false
+		}
+		_, ok := forceAnalyzerTags[oper]
+		return !ok
+	} else {
+		return true
+	}
+}
+
+func OnAppDeliverTxEnter() {
+	if analyzer != nil {
+		analyzer.onAppDeliverTxEnter()
+	}
+}
+
+func OnCommitDone() {
+	if analyzer != nil {
+		analyzer.onCommitDone()
+	}
+}
+
+func StartTxLog(oper string) {
+	if !skip(analyzer, oper) {
+		analyzer.startTxLog(oper)
+	}
+}
+
+func StopTxLog(oper string) {
+	if !skip(analyzer, oper) {
+		analyzer.stopTxLog(oper)
+	}
+}
+
+
+func SetDynamicConfig(c IDynamicConfig) {
+	dynamicConfig = c
+}
+
+type IDynamicConfig interface {
+	GetEnableAnalyzer() bool
+}
+
+type MockDynamicConfig struct {
+}
+
+func (c MockDynamicConfig) GetEnableAnalyzer() bool {
+	return viper.GetBool(FlagEnableAnalyzer)
 }
